+++ tests/feature/session_functions/src/rflx-test-session_functions.ads
@@ -6,8 +6,4 @@ is
 
    type Context is null record;
 
-   function Initialize return Context;
-
-   procedure Finalize (Ctx : in out Context);
-
 end RFLX.Test.Session_Functions;